Fixed 1 bug and got rid of nodeTransitions (was quite stupid from the beginning)
- When Undoing and Redoing inside an element that's inside another element the connections outside of it got disconnected

+++ TFG/GUI Control/BehaviourTree.cs
@@ -160,18 +160,21 @@ public class BehaviourTree : ClickableElement
     /// Deletes the <paramref name="node"/> and its children
     /// </summary>
     /// <param name="node"></param>
-    public void DeleteNode(BehaviourNode node)
+    public void DeleteNode(BehaviourNode node, bool deleteTransitions = true)
     {
         if (nodes.Remove(node))
         {
-            foreach (TransitionGUI transition in connections.FindAll(t => node.Equals(t.fromNode)))
+            if (deleteTransitions)
             {
-                connections.Remove(transition);
-                DeleteNode((BehaviourNode)transition.toNode);
-            }
-            foreach (TransitionGUI transition in connections.FindAll(t => node.Equals(t.toNode)))
-            {
-                connections.Remove(transition);
+                foreach (TransitionGUI transition in connections.FindAll(t => node.Equals(t.fromNode)))
+                {
+                    connections.Remove(transition);
+                    DeleteNode((BehaviourNode)transition.toNode);
+                }
+                foreach (TransitionGUI transition in connections.FindAll(t => node.Equals(t.toNode)))
+                {
+                    connections.Remove(transition);
+                }
             }
 
             if (node.subElem == null)
